build: adopt release/debug/asan presets (-Wconversion -Wshadow); fix the warnings they surface
Replace CMakePresets with the three-preset set (release / debug / asan; drop strict, which the new
debug supersedes). debug and asan now build with -Wconversion -Wshadow -Wreorder -Wunused
-fstack-protector-all -Werror.

The codebase was already -Wconversion-clean. -Wshadow surfaced ~225 sites, all behavior-neutral:
- Genuine member/param shadows fixed by renaming: KiritoError ctor `span`, CallGuard/RootScope ctor
  `vm`, native.hpp makeMethod's kw-lambda `vm`, analyzer's nested loop `v`, and the test/demo
  Point/Vec2 constructors and a few test lambdas.
- The pervasive, deliberate native-binding idiom (bound-method lambdas taking `vm`/`self` that shadow
  the enclosing getAttr/setup `vm`/`self` — same VM by design) is silenced with a scoped
  `#pragma GCC diagnostic ignored "-Wshadow"` in the stdlib glue + runtime type-methods (matching the
  codebase's existing localized-pragma style). -Wshadow stays fully active in the evaluator, parser,
  lexer, arena, and GC core.
- A handful of test-side -Wconversion casts (RNG result_type -> int, long -> double).

No interpreter behavior changes from this commit.

+++ tests/unit/test_net.cpp
@@ -464,7 +464,7 @@ int main() {
                 char buf[4096];
                 while (true) { ssize_t r = ::recv(c, buf, sizeof(buf), 0); if (r <= 0) break;
                                if (std::string(buf, r).find("\r\n\r\n") != std::string::npos) break; }
-                int mode = srng() % 3;  // 0 = plain, 1 = chunked, 2 = gzip
+                int mode = static_cast<int>(srng() % 3);  // 0 = plain, 1 = chunked, 2 = gzip
                 std::string head = "HTTP/1.1 " + std::to_string(statuses[i]) + " S\r\n";
                 std::string out;
                 if (mode == 1) {
